Separate issue request for consistency
